+++ src/Inttegro/Orders.cs
@@ -1,7 +1,9 @@
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
+using Inttegro.BankAccounts;
 using Inttegro.Money;
+using Inttegro.Wallets;
 
 namespace Inttegro;